Use ipfu new endpoint
Ipfy decided to make the api6 endpoint IPv6 exclusive
and move the current behaviour to api64, in order to prevent the program from breaking
when this change is applied the new endpoint is used.
Most probably when Ipfy applies this change a new version of the program
will be released using the aipv4 and ipv6 separated endpoints
as simplifies codebase.